refactor to avoid cyclic module dependency
